+++ src/lib/appliances/appliances.table.tsx
@@ -1,14 +1,12 @@
 import React, { useMemo, useState } from "react";
 import {
   Card,
-  Accordion,
-  Button,
   Container,
-  Modal,
-  Form,
+  Button,
 } from "react-bootstrap";
 import DataTable, { TableColumn } from "react-data-table-component";
 import { DATA_TABLE_DEFAULT_STYLE } from "../constants";
+import AddAppliance from "./app.appliance";
 
 interface ApplianceGroup {
   name: string;
@@ -21,14 +19,7 @@ const AppliancesTable: React.FC<DataTableProps<Appliance>> = ({
 }) => {
   const [data, setData] = useState<Appliance[]>(initialData);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-  const [currentAppliance, setCurrentAppliance] = useState<Appliance | null>(
-    null
-  );
-  const [step, setStep] = useState<number>(1);
-  const [applianceName, setApplianceName] = useState<string>("");
-  const [applianceDetails, setApplianceDetails] = useState<Partial<Appliance>>(
-    {}
-  );
+  const [currentAppliance, setCurrentAppliance] = useState<Appliance | null>(null);
 
   const handleShow = () => setIsModalOpen(true);
   const handleClose = () => {
@@ -38,51 +29,20 @@ const AppliancesTable: React.FC<DataTableProps<Appliance>> = ({
 
   const resetForm = () => {
     setCurrentAppliance(null);
-    setStep(1);
-    setApplianceName("");
-    setApplianceDetails({});
   };
 
-  const handleNextStep = () => {
-    if (step === 1) {
-      setStep(2);
-      if (currentAppliance) {
-        setApplianceName(currentAppliance.name);
-        setApplianceDetails({
-          brand: currentAppliance.brand,
-          type: currentAppliance.type,
-          additionalInfo: currentAppliance.additionalInfo,
-          referenceUrl: currentAppliance.referenceUrl,
-        });
-      }
-    } else {
-      addAppliance();
-    }
-  };
-
-  const addAppliance = () => {
-    const newAppliance: Appliance = {
-      name: applianceName,
-      brand: applianceDetails.brand || "",
-      type: applianceDetails.type || "",
-      additionalInfo: applianceDetails.additionalInfo || "",
-      referenceUrl: applianceDetails.referenceUrl || "",
-    };
-
+  const addAppliance = (newAppliance: Appliance) => {
     setData((prevData) => {
-      // Check if we're updating an existing appliance
       const existingApplianceIndex = prevData.findIndex(
         (appliance) => appliance.name === currentAppliance?.name
       );
 
       if (existingApplianceIndex !== -1) {
-        // Update existing appliance
-        const updatedData = [...prevData]; // Create a new array
-        updatedData[existingApplianceIndex] = newAppliance; // Update the specific appliance
-        return updatedData; // Return the updated array
+        const updatedData = [...prevData]; 
+        updatedData[existingApplianceIndex] = newAppliance; 
+        return updatedData; 
       } else {
-        // Add new appliance
-        return [...prevData, newAppliance]; // Return the new array with the new appliance
+        return [...prevData, newAppliance]; 
       }
     });
 
@@ -115,14 +75,14 @@ const AppliancesTable: React.FC<DataTableProps<Appliance>> = ({
           details: [],
         };
       }
-      applianceMap[appliance.name].quantity += 1; // Increment quantity for each occurrence
+      applianceMap[appliance.name].quantity += 1; 
       applianceMap[appliance.name].details.push(
         `${appliance.type || "N/A"} (${appliance.brand || "N/A"})`
       );
     });
 
     return Object.values(applianceMap);
-  }, [data]); // Recalculate groupedData whenever data changes
+  }, [data]);
 
   const columns: TableColumn<ApplianceGroup>[] = useMemo(() => {
     return [
@@ -132,19 +92,9 @@ const AppliancesTable: React.FC<DataTableProps<Appliance>> = ({
         sortable: true,
         cell: (row) => (
           <Container fluid className="d-flex w-100">
-            <Accordion className="flex-grow-1 bg-transparent mx-4">
-              <Accordion.Item
-                eventKey={row.name}
-                className="custom-accordion-item"
-              >
-                <Accordion.Header className="d-flex justify-content-between align-items-center w-100">
-                  <div className="flex-grow-1">
-                    {row.name} (Quantity: {row.quantity})
-                  </div>
-                </Accordion.Header>
-                <Accordion.Body>{row.details.join(", ")}</Accordion.Body>
-              </Accordion.Item>
-            </Accordion>
+            <div className="flex-grow-1 mx-4">
+              {row.name} (Quantity: {row.quantity})
+            </div>
             <div className="d-flex justify-content-center align-items-center">
               <Button
                 size="sm"
@@ -179,113 +129,16 @@ const AppliancesTable: React.FC<DataTableProps<Appliance>> = ({
           <DataTable
             columns={columns}
             data={groupedData}
-            striped
-            highlightOnHover
-            pagination
             customStyles={DATA_TABLE_DEFAULT_STYLE}
           />
         </Card.Body>
       </Card>
-      <Modal show={isModalOpen} onHide={handleClose} centered>
-        <Modal.Header closeButton>
-          <Modal.Title>
-            {step === 1 ? "Select Appliance" : "Enter Details"}
-          </Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {step === 1 ? (
-            <Form>
-              <Form.Group controlId="applianceName">
-                <Form.Label>Appliance Name</Form.Label>
-                <Form.Control
-                  as="select"
-                  value={applianceName}
-                  onChange={(e) => setApplianceName(e.target.value)}
-                >
-                  <option value="">Select...</option>
-                  <option value="fridge">Fridge</option>
-                  <option value="extractor">Extractor</option>
-                  <option value="tap">Tap</option>
-                  <option value="oven">Oven</option>
-                  <option value="dishwasher">Dishwasher</option>
-                  <option value="hob">Hob</option>
-                </Form.Control>
-              </Form.Group>
-            </Form>
-          ) : (
-            <Form>
-              <Form.Group controlId="brand">
-                <Form.Label>Brand</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter brand"
-                  value={applianceDetails.brand || ""}
-                  onChange={(e) =>
-                    setApplianceDetails({
-                      ...applianceDetails,
-                      brand: e.target.value,
-                    })
-                  }
-                />
-              </Form.Group>
-              <Form.Group controlId="type">
-                <Form.Label>Type</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter type"
-                  value={applianceDetails.type || ""}
-                  onChange={(e) =>
-                    setApplianceDetails({
-                      ...applianceDetails,
-                      type: e.target.value,
-                    })
-                  }
-                />
-              </Form.Group>
-              <Form.Group controlId="additionalInfo">
-                <Form.Label>Additional Info</Form.Label>
-                <Form.Control
-                  type="text"
-                  placeholder="Enter additional info"
-                  value={applianceDetails.additionalInfo || ""}
-                  onChange={(e) =>
-                    setApplianceDetails({
-                      ...applianceDetails,
-                      additionalInfo: e.target.value,
-                    })
-                  }
-                />
-              </Form.Group>
-              <Form.Group controlId="referenceUrl">
-                <Form.Label>Reference URL</Form.Label>
-                <Form.Control
-                  type="url"
-                  placeholder="Enter reference URL"
-                  value={applianceDetails.referenceUrl || ""}
-                  onChange={(e) =>
-                    setApplianceDetails({
-                      ...applianceDetails,
-                      referenceUrl: e.target.value,
-                    })
-                  }
-                />
-              </Form.Group>
-            </Form>
-          )}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={handleClose}>
-            Cancel
-          </Button>
-          <Button variant="primary" onClick={handleNextStep}>
-            {step === 1
-              ? "Next"
-              : currentAppliance
-              ? "Update Appliance"
-              : "Add Appliance"}
-          </Button>
-        </Modal.Footer>
-      </Modal>
+      <AddAppliance
+        show={isModalOpen}
+        handleClose={handleClose}
+        addAppliance={addAppliance}
+        currentAppliance={currentAppliance}
+      />
     </Container>
   );
 };
