+++ src/pages/appliances.page.tsx
@@ -1,14 +1,14 @@
-import React from 'react'
-import ProfileProjectLayout from '../lib/components/profile.project/profile.project.layout'
-import AppliancesTable from '../lib/appliances/appliances.table'
-import { DUMMY_APPLIANCES } from '../lib/data/dummy_appliances'
+import React from "react";
+import ProfileProjectLayout from "../lib/components/profile.project/profile.project.layout";
+import AppliancesTable from "../lib/appliances/appliances.table";
+import { DUMMY_APPLIANCES } from "../lib/data/dummy_appliances";
 
 const AppliancesPage = () => {
   return (
-    <ProfileProjectLayout>
+      <ProfileProjectLayout>
         <AppliancesTable initialData={DUMMY_APPLIANCES}/>
-    </ProfileProjectLayout>
-  )
-}
+      </ProfileProjectLayout>
+  );
+};
 
-export default AppliancesPage
+export default AppliancesPage;
